Update to newest framework patterns
I was able to get the cluster up and running with these changes. Haven't
tested workloads, but all system pods appear to be running fine.
