+++ .gitignore
@@ -1,0 +1,2 @@
+lib/ops
+lib/ops-*.dist-info
