+++ hooks/install
@@ -1,1 +1,1 @@
-../lib/juju/main.py
+../lib/charm.py
