--- lib/test/bin/relation-list
@@ -1,16 +1,0 @@
-#!/bin/bash
-
-fail_not_found() {
-    1>&2 echo "ERROR invalid value \"$1\" for option -r: relation not found"
-    exit 2
-}
-
-case $2 in
-    1) echo '["remote/0"]' ;;
-    2) echo '["remote/0"]' ;;
-    3) fail_not_found $2 ;;
-    4) echo '["remoteapp1/0"]' ;;
-    5) echo '["remoteapp1/0"]' ;;
-    6) echo '["remoteapp2/0"]' ;;
-    *) fail_not_found $2 ;;
-esac
