--- lib/test/charms/test_main/config.yaml
@@ -1,1 +1,0 @@
-"options": {}
