--- lib/test/charms/test_main/hooks/disks-storage-attached
@@ -1,1 +1,0 @@
-install
