--- lib/test/charms/test_main/hooks/install
@@ -1,1 +1,0 @@
-../lib/juju/main.py
