--- lib/test/charms/test_main/lib/juju
@@ -1,1 +1,0 @@
-../../../../juju
